IBVH.jl Documentation

BVH Construction & Traversal

IBVH.BVH — Type

struct BVH{VN<:(AbstractVector), VL<:(AbstractVector), VO<:(AbstractVector)}

Implicit bounding volume hierarchy constructed from an iterable of some geometric primitives' (e.g. triangles in a mesh) bounding volumes forming the ImplicitTree leaves. The leaves and merged nodes above them can have different types - e.g. BSphere{Float64} for leaves merged into larger BBox{Float64}.

The initial geometric primitives are sorted according to their Morton-encoded coordinates; the unsigned integer type used for the Morton encoding can be chosen between Union{UInt16, UInt32, UInt64}.

Finally, the tree can be incompletely-built up to a given built_level and later start contact detection downwards from this level, e.g.:

Implicit tree from 5 bounding volumes - i.e. the real leaves

Tree Level          Nodes & Leaves               Build Up    Traverse Down
    1                     1                         Ʌ              |
    2             2               3                 |              |
    3         4       5       6        7v           |              |
    4       8   9   10 11   12 13v  14v  15v        |              V
            -------Real------- ---Virtual---

Methods

function BVH(
    bounding_volumes::AbstractVector{L},
    node_type::Type{N}=L,
    morton_type::Type{U}=UInt,
    built_level::Integer=1,
) where {L, N, U <: MortonUnsigned}

Fields

tree::ImplicitTree{Int}
nodes::VN   <: AbstractVector
leaves::VL  <: AbstractVector
order::VO   <: AbstractVector
built_level::Int
stats::BVHStats

Examples

Simple usage with bounding spheres and default 64-bit types:

using IBVH
using IBVH: BSphere
using StaticArrays

# Generate some simple bounding spheres
bounding_spheres = [
    BSphere(SA[0., 0., 0.], 0.5),
    BSphere(SA[0., 0., 1.], 0.6),
    BSphere(SA[0., 0., 2.], 0.5),
    BSphere(SA[0., 0., 3.], 0.4),
    BSphere(SA[0., 0., 4.], 0.6),
]

# Build BVH
bvh = BVH(bounding_spheres)

# Traverse BVH for contact detection
traversal = traverse(bvh)
@show traversal.contacts;
;

# output
traversal.contacts = [(4, 5), (1, 2), (2, 3)]

Using Float32 bounding spheres for leaves, Float32 bounding boxes for nodes above, and UInt32 Morton codes:

using IBVH
using IBVH: BBox, BSphere
using StaticArrays

# Generate some simple bounding spheres
bounding_spheres = [
    BSphere{Float32}(SA[0., 0., 0.], 0.5),
    BSphere{Float32}(SA[0., 0., 1.], 0.6),
    BSphere{Float32}(SA[0., 0., 2.], 0.5),
    BSphere{Float32}(SA[0., 0., 3.], 0.4),
    BSphere{Float32}(SA[0., 0., 4.], 0.6),
]

# Build BVH
bvh = BVH(bounding_spheres, BBox{Float32}, UInt32)

# Traverse BVH for contact detection
traversal = traverse(bvh)
@show traversal.contacts;
;

# output
traversal.contacts = [(4, 5), (1, 2), (2, 3)]

Build BVH up to level 2 and start traversing down from level 3, reusing the previous traversal cache:

bvh = BVH(bounding_spheres, BBox{Float32}, UInt32, 2)
traversal = traverse(bvh, 3, traversal)

IBVH.traverse — Function

traverse(
    bvh::BVH,
    start_level=max(bvh.tree.levels ÷ 2, bvh.built_level),
    cache::Union{Nothing, BVHTraversal}=nothing,
)::BVHTraversal

Traverse bvh downwards from start_level, returning all contacting bounding volume leaves. The returned BVHTraversal also contains two contact buffers that can be reused on future traversals.

Examples

using IBVH
using IBVH: BBox, BSphere
using StaticArrays

# Generate some simple bounding spheres
bounding_spheres = [
    BSphere{Float32}(SA[0., 0., 0.], 0.5),
    BSphere{Float32}(SA[0., 0., 1.], 0.6),
    BSphere{Float32}(SA[0., 0., 2.], 0.5),
    BSphere{Float32}(SA[0., 0., 3.], 0.4),
    BSphere{Float32}(SA[0., 0., 4.], 0.6),
]

# Build BVH
bvh = BVH(bounding_spheres, BBox{Float32}, UInt32)

# Traverse BVH for contact detection
traversal = traverse(bvh, 2)

# Reuse traversal buffers for future contact detection - possibly with different BVHs
traversal = traverse(bvh, 2, traversal)
@show traversal.contacts;
;

# output
traversal.contacts = [(4, 5), (1, 2), (2, 3)]

Other BVH Interfaces

IBVH.BVHTraversal — Type

struct BVHTraversal{VC<:(AbstractVector)}

Collected BVH traversal contacts list, plus the two buffers cache1 and cache2 which can be reused for future traversals to minimise memory allocations.

IBVH.BVHStats — Type

mutable struct BVHStats

Collected statistics about a BVH construction and contact traversal; populated after calling traverse.

Fields

start_level::Union{Nothing, Int64}: Default: nothing

num_checks::Union{Nothing, Int64}: Default: nothing

num_contacts::Union{Nothing, Int64}: Default: nothing

IBVH.IndexPair — Type

struct Tuple{Int64, Int64}

Alias for a tuple of two indices representing e.g. a contacting pair.

Index

IBVH.BVH

IBVH.BVHStats

IBVH.BVHTraversal

IBVH.ImplicitTree

IBVH.IndexPair

IBVH.MortonUnsigned

IBVH.TaskPartitioner

IBVH.bounding_volumes_extrema

IBVH.isvirtual

IBVH.level_indices

IBVH.memory_index

IBVH.morton_encode

IBVH.morton_encode!

IBVH.morton_encode_single

IBVH.morton_scaling

IBVH.morton_split3

IBVH.traverse