Morton Encoding

IBVH.MortonUnsigned — Type

Acceptable unsigned integer types for Morton encoding.

IBVH.MortonUnsignedType — Type

Type values of acceptable unsigned integer types for Morton encoding.

IBVH.morton_encode! — Function

morton_encode(bounding_volumes, U::MortonUnsignedType=UInt32)
morton_encode!(mortons, bounding_volumes, U)
morton_encode!(mortons, bounding_volumes, mins, maxs, U)

Encode each bounding volume into Vector{U} of corresponding Morton codes such that they uniformly cover the maximum Morton range given an unsigned integer type U.

If not provided, allocate Vector{U} for Morton codes. Likewise, compute dimension-wise exclusive minima and maxima if not provided.

IBVH.morton_encode_single — Function

morton_encode_single(centre, mins, maxs, U::MortonUnsignedType=UInt32)

Return Morton code for a single 3D position centre scaled uniformly between mins and maxs. Works transparently for SVector, Vector, etc. with eltype UInt16, UInt32 or UInt64.

IBVH.morton_scaling — Function

morton_scaling(::Type{UInt16}) = 2^5
morton_scaling(::Type{UInt32}) = 2^10
morton_scaling(::Type{UInt64}) = 2^21

Exclusive maximum number possible to use for 3D Morton encoding for each type.

IBVH.morton_split3 — Function

morton_split3(v::UInt16)
morton_split3(v::UInt32)
morton_split3(v::UInt64)

Shift a number's individual bits such that they have two zeros between them.

IBVH.bounding_volumes_extrema — Function

bounding_volumes_extrema(bounding_volumes)

Compute exclusive lower and upper bounds in iterable of bounding volumes, e.g. Vector{BBox}.